Re-organize asset file structure
