+++ src/assets/readme/model-switching.png
