+++ src/assets/services/OpenAI.png
